requirements.txt
connexion[swagger-ui] >= 2.6.0; python_version>="3.6"
Flask>=2.1.1
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
PyMySQL>=1.1.1
DBUtils>=3.1.0
cryptography
python-decouple

-r swagger_server/requirements.txt